WIP: OcdFileImport/Export: Maintain object gap points
For the sake of enhanced OCD format compatibility, store the various gap
point types in the OMAP file and save it again on OCD export. Mapper
does not make use of the no-left, no-right and gap information bits.

The use of MapCoord::Flag::GapPoint for line object gaps is not exactly
the right choice. While it serves the purpose of maintaining the
information bit, the object rendering is not correct.