Map: Rename color manipulation methods
The number accepted by getMapColor() and similar methods is not exactly
an index. The value is treated as the color rendering priority, it is
synced into the color Priority member and should be treated as such.
Namely, reordering the color table changes the priorities
invalidating any previous parameter values. This commit makes it obvious
what value we are using for color lookup.