OcdFileImport/Export: Persist OCD color numbers
Use the color id to store the OCD color number. We know that the
registration black may move in the table, get a new number, or might be
dropped. The thing is that Mapper has registration black as a special
color that cannot get an OCD color number, and we merely guess whether
we need the color in the table in the export path.

Closes GH-1245.