GeoreferencingDialog: Fix declination lookup
The declination lookup at www.ngdc.noaa.gov now requests an additional
parameter called 'key'. We are using a default key but in case the
lookup fails, we give the user an opportunity to enter an alternate key.
If the new key works, we store it in the program settings.

Based on work by Matthias Kuehlewein.