VirtualPath: Export tangent from findClosestPointTo()
The tangent vector is essential for the move parallel tool where we
want to calculate the distance of the cursor from the tangent line
touching the path in question. Please note that we are generating a
"pseudo tangent" for corners and line ends to enable sensible
behavior during mouse drag events.