MoveParallelTool: Add the tool for parallel object moves
The tool moves lines and area boundaries to create a larger or
a smaller shape parallel to the original object. The tool reuses
LineSymbol::shiftCoordinates() which currently does not handle shape
degeneration. I.e., curve turns do not shrink into points but rather
form a "negative" shape that is a mathematically exact image of the
original curve but usually is not what the user expects.