LibreMapper: Switch to Qt6
This Mapper state does not run as IconScalingEngine takes the process
down through infinite recursion. The fix is in a separate patch.

We are keeping Qt5Compat in place to maintain access to QTextCodec. Qt6
does not have a full-scale replacement for the character encoding
facility at the moment.

The std::iterator deprecation warning will be tackled later.