OcdFileImport/Export: Fix registration black id assignment
Registration black may not get its original id even if the id is free
after re-export to .ocd.  The thing is that spot colors were assigned
new id's in the OCD import path.  The export path saw the low id numbers
as used and searched for a new id for the registration black.

This patch fixes the issue by assigning high id's to the spot colors so
that the low numbers remain available for the registration black id
selection.

Amends 5e98f2f ("OcdFileImport/Export: Persist OCD color numbers").