ColorListWidget: Warn about spot color removal
A color entry may serve as a spot color without direct use in symbols.
So far Mapper allowed silent removal of these color entries. This patch
adds a warning which is shown when the user tries to remove such a spot
color.

Co-authored-by: Matthias Kuehlewein <github@kuehlewein.de>
Co-authored-by: Libor Pechacek <lpechacek@gmx.com>

[lpechacek] Contains a revert of 0d46a1a ("ColorListWidget: Warn about
spot color removal") which is the first implementation of this change.