MapEditorController: Disable overprinting when deleting the last spot color
When overprinting simulation is enabled while deleting the last spot
color, the map editor will become empty, and the overprinting action
becomes disabled. However, the program didn't actually activate the
overprinting simulation mode, so the user was locked out from
re-enabling the normal rendering of the map. We must explicitly trigger
the action.

(Reactions are bound to trigger() to avoid hard-to-debug activation
loops from normal state changes.)

Co-authored-by: Libor Pechacek <lpechacek@gmx.com>

[lpechacek] Contains a revert of db060ff ("MapEditorController: Disable
overprinting when spot colors are gone") which is the first
implementation of this change.