ISOM 2017-2: Update to Revision 6 (colors)
Colors renamed and reordered in line with Printing And Colour
Definitions. Colors that are not defined in the standard are
marked with "(extra)" suffix in their name.

Color knockout property is optimized for OCD file export. Mind that OCD
overprint applies to CMYK process and *not* spot colors. All colors,
except full black, are now set to "knockout" which translates to OCD "no
overprint". That aligns the symbol set in exported files with the usual
symbol sets used in O-maps outside Mapper realm. Of course, it breaks
spot color rendering in Mapper. We are giving priority to compatibility
over rendering at the moment.

At the same time we are adjusting the separation list for individual
colors and set zero percent where the color should erase the underlying
spot color rendering. This is the standard method for knockout in the
OCD format representation.