Map information tool: Information about map properties
The map information tool informs about the number of map objects,
templates and undo-/redo-steps.
It informs about the number and names of map parts, symbols, colors,
and fonts.
It shows for each color the symbol(s) using the color.
For each symbol in use the tool shows the used color(s) and the number
of related objects.
The report can be exported as text file.

[lpechacek] Port to LibreMapper.