WIP: More realistic map color rendering
This patch takes the on-screen map color appearance closer to what a PDF
viewever renders from a CMYK-enabled export.

As the difference to the previous state is substantial, the user should
be able to choose between the old rendering and the new one. Possibly,
the OCAD-like rendering can be added to the mix if we derive the formula
from rendering tests.