Fix rendering of cross-hatched areas (GH-2102)
The assumption that the first point lies in the top left corner of the
bounding box is not valid. As a result, the rendering of cross-hatched
area patterns was broken when the area extended beyond the viewport.

Amends commit 3590c35e80739 (LineRenderable: Fix extent calculation in
simple constructor).
Closes GH-2101 (Incorrect rendering of area cross-hatch pattern).

Co-authored-by: Kai Pastor <dg0yt@darc.de>

[lpechacek] This reverts c3f86a75934a ("LineRenderable: Fix rendering of
cross-hatched areas") which was the previous implementation.