Minimum area and length in combined symbols (GH-2309)
Combined symbols may contain subsymbols with a minimum area size
or a minimum line length. This change makes these properties
available to the MeasureWidget. It is based on generalizing these
read-only properties into the Symbols interface, and adding
general (Symbol) and specific (CombinedSymbol) implementations.

Closes GH-2083.

---------

Co-authored-by: Kai Pastor <dg0yt@darc.de>

[lpechacek] This reverts commit 676f6cb1844b ("MeasureWidget: Warn about
min size in combined symbols") which was the LibreMapper implementation.