MapFindFeature: Exclude hidden/protected objects (GH-2371)
The 'Find next' search failed if the object query returned an object
with either a hidden or protected symbol since these kind of objects
can't be added to an object selection. This change excludes such
objects from the result set. A test is added.

---------

Co-authored-by: Kai Pastor <dg0yt@darc.de>
Ported-to-LibreMapper: Libor Pechacek <lpechacek@gmx.com>