OgrFileExport: Exclude objects with hidden or helper symbols
The geospatial vector export shall not export objects with hidden or
helper symbols. This change restructures the export code so that the
filtering is also applied when the "Create a layer for each symbol"
setting in the GDAL/OGR configuration was is not set then objects.

Ported-to-LibreMapper: Libor Pechacek <lpechacek@gmx.com>