TemplateListWidget: Consider cancellation of template georeferencing
If the user activates georeferencing for a template where
georeferencing was switched off, a dialog will show up to select the
coordinate reference system. If the dialog was cancelled the
'Georeferenced' checkmark would be set and 'Positioning' would be
discarded as if coordinate reference system selection succeeded.

Ported-to-LibreMapper: Libor Pechacek <lpechacek@gmx.com>